feat(core): allow not to add render hook

+++ packages/core/src/module/app.ts
@@ -390,22 +390,24 @@ export class App {
   // Calls to render do compatible with two different sandbox
   private callRender(provider: interfaces.Provider, isMount: boolean) {
     const { appInfo, rootElement } = this;
-    provider &&
+    if (provider && provider.render) {
       provider.render({
         dom: rootElement,
         basename: appInfo.basename,
         appRenderInfo: { isMount },
       });
+    }
   }
 
   // Call to destroy do compatible with two different sandbox
   private callDestroy(provider: interfaces.Provider, isUnmount: boolean) {
     const { rootElement, appContainer } = this;
-    provider &&
+    if (provider && provider.destroy) {
       provider.destroy({
         dom: rootElement,
         appRenderInfo: { isUnmount },
       });
+    }
     this.entryManager.DOMApis.removeElement(appContainer);
   }
